requirements.txt
Django>=5.0,<6
djangorestframework>=3.14,<4
djangorestframework-simplejwt>=5.3,<6
Pillow>=10.0,<11
